Broke up some functionality of data_store into smaller parts
